backend/requirements.txt
Django>=5.0
django-cors-headers>=4.0
python-dotenv>=1.0
